backend/app/app/api/api_v1/endpoints/items_gql.py





import strawberry

from fastapi import FastAPI
from strawberry.fastapi import GraphQLRouter

@strawberry.type
class Query:
    @strawberry.field
    def hello(self) -> str:
        return "Hello World"


schema = strawberry.Schema(Query)

router = GraphQLRouter(schema)




